+++ app/_components/credits/CreditsDialog.tsx
@@ -6,8 +6,8 @@ import React from "react";
 
 interface LogoAttribution {
   name: string;
-  license: string;
-  licenseUrl: string;
+  license?: string;
+  licenseUrl?: string;
   source: string;
   sourceUrl: string;
 }
@@ -24,6 +24,13 @@ const CreditsDialog: React.FC = function () {
       sourceUrl: 'https://github.com/google/material-design-icons'
     },
     {
+      name: 'School',
+      license: 'Apache 2.0',
+      licenseUrl: 'https://github.com/google/material-design-icons/blob/master/LICENSE',
+      source: 'Google',
+      sourceUrl: 'https://github.com/google/material-design-icons'
+    },
+    {
       name: 'Accenture',
       license: 'Creative Commons Zero v1.0 Universal',
       licenseUrl: 'https://github.com/simple-icons/simple-icons/blob/develop/LICENSE.md',
@@ -32,17 +39,13 @@ const CreditsDialog: React.FC = function () {
     },
     {
       name: 'Bootstrap',
-      license: 'CC0 1.0 Universal',
-      licenseUrl: 'https://raw.githubusercontent.com/gilbarbara/logos/master/LICENSE.txt',
-      source: 'Gil Barbara',
-      sourceUrl: 'https://github.com/gilbarbara/logos'
+      source: 'Bootstrap',
+      sourceUrl: 'https://getbootstrap.com/docs/4.0/about/brand'
     },
     {
       name: 'ChatGPT',
-      license: 'CC BY-SA 4.0',
-      licenseUrl: 'https://creativecommons.org/licenses/by-sa/4.0/',
-      source: 'Arcticons',
-      sourceUrl: 'https://github.com/Arcticons-Team/Arcticons'
+      source: 'OpenAI',
+      sourceUrl: 'https://openai.com/brand/'
     },
     {
       name: 'Confluence',
@@ -61,9 +64,9 @@ const CreditsDialog: React.FC = function () {
     {
       name: 'Docker',
       license: 'CC0 1.0 Universal',
-      licenseUrl: 'https://raw.githubusercontent.com/gilbarbara/logos/master/LICENSE.txt',
-      source: 'Gil Barbara',
-      sourceUrl: 'https://github.com/gilbarbara/logos'
+      licenseUrl: 'https://www.docker.com/legal/trademark-guidelines/',
+      source: 'Docker',
+      sourceUrl: 'https://www.docker.com/company/newsroom/media-resources/'
     },
     {
       name: 'edX',
@@ -81,17 +84,15 @@ const CreditsDialog: React.FC = function () {
     },
     {
       name: 'Git',
-      license: 'Apache 2.0',
-      licenseUrl: 'https://github.com/Templarian/MaterialDesign/blob/master/LICENSE',
-      source: 'Pictogrammers',
-      sourceUrl: 'https://github.com/Templarian/MaterialDesign'
+      license: 'Creative Commons Attribution 3.0 Unported License',
+      licenseUrl: 'https://git-scm.com/downloads/logos',
+      source: 'Git',
+      sourceUrl: 'https://git-scm.com/downloads/logos'
     },
     {
       name: 'Github',
-      license: 'Apache 2.0',
-      licenseUrl: 'https://github.com/Templarian/MaterialDesign/blob/master/LICENSE',
-      source: 'Pictogrammers',
-      sourceUrl: 'https://github.com/Templarian/MaterialDesign'
+      source: 'Github',
+      sourceUrl: 'https://github.com/logos'
     },
     {
       name: 'JavaScript',
@@ -130,10 +131,8 @@ const CreditsDialog: React.FC = function () {
     },
     {
       name: 'LinkedIn',
-      license: 'Apache 2.0',
-      licenseUrl: 'https://github.com/Templarian/MaterialDesign/blob/master/LICENSE',
-      source: 'Pictogrammers',
-      sourceUrl: 'https://github.com/Templarian/MaterialDesign'
+      source: 'LinkedIN',
+      sourceUrl: 'https://brand.linkedin.com/downloads'
     },
     {
       name: 'MySQL',
@@ -151,10 +150,8 @@ const CreditsDialog: React.FC = function () {
     },
     {
       name: 'NodeJS',
-      license: 'CC0 1.0 Universal',
-      licenseUrl: 'https://raw.githubusercontent.com/gilbarbara/logos/master/LICENSE.txt',
-      source: 'Gil Barbara',
-      sourceUrl: 'https://github.com/gilbarbara/logos'
+      source: 'NodeJS',
+      sourceUrl: 'https://nodejs.org/en/about/branding'
     },
     {
       name: 'PHP',
@@ -176,6 +173,13 @@ const CreditsDialog: React.FC = function () {
       licenseUrl: 'https://raw.githubusercontent.com/gilbarbara/logos/master/LICENSE.txt',
       source: 'Gil Barbara',
       sourceUrl: 'https://github.com/gilbarbara/logos'
+    },
+    {
+      name: 'SonarQube',
+      license: 'Creative Commons Zero v1.0 Universal',
+      licenseUrl: 'https://github.com/simple-icons/simple-icons/blob/develop/LICENSE.md',
+      source: 'Simple Icons Collaborators',
+      sourceUrl: 'https://github.com/simple-icons/simple-icons'
     },
     {
       name: 'UI5',
@@ -228,6 +232,21 @@ const CreditsDialog: React.FC = function () {
         centered
       >
         <Stack pr="xs">
+          <Text fw={700}>Frameworks and Libraries</Text>
+          <List>
+            <List.Item key="mantine">
+              Framework by <Anchor href="https://nextjs.org/" target="_blank">Next.js</Anchor> licensed under <Anchor href="https://github.com/vercel/next.js/blob/main/license.md" target="_blank">MIT License</Anchor>.
+            </List.Item>
+            <List.Item key="mantine">
+              UI Library by <Anchor href="https://mantine.dev/" target="_blank">Mantine</Anchor> licensed under <Anchor href="https://github.com/mantinedev/mantine/blob/master/LICENSE" target="_blank">MIT License</Anchor>.
+            </List.Item>
+            <List.Item key="mantine">
+              Typing animation library by <Anchor href="https://github.com/mattboldt/typed.js" target="_blank">Matt Boldt</Anchor> licensed under <Anchor href="https://github.com/mattboldt/typed.js#MIT-1-ov-file" target="_blank">MIT License</Anchor>.
+            </List.Item>
+            <List.Item key="mantine">
+              Icon component library by <Anchor href="https://tabler.io/" target="_blank">Tabler.io</Anchor> licensed under <Anchor href="https://github.com/tabler/tabler?tab=MIT-1-ov-file#readme" target="_blank">MIT License</Anchor>.
+            </List.Item>
+          </List>
           <Text fw={700}>Logo Attributions</Text>
           <List>
             {
@@ -247,6 +266,12 @@ const CreditsDialog: React.FC = function () {
 }
 
 const LogoAttributionComponent: React.FC<LogoAttribution> = function (props) {
+  if (!props.license || !props.licenseUrl) {
+    return <Text>
+      {props.name} icon by <Anchor href={props.sourceUrl} target="_blank">{props.source}</Anchor>.
+    </Text>
+  }
+
   return (
     <Text>
       {props.name} icon by <Anchor href={props.sourceUrl} target="_blank">{props.source}</Anchor> licensed under <Anchor href={props.licenseUrl} target="_blank">{props.license || 'a custom license'}</Anchor>.