+++ app/_sections/skills/SkillsSection.tsx
@@ -112,7 +112,12 @@ const SkillsSection: React.FC = function () {
       description: 'test',
     },
     {
-      logo: '/logos/chatgpt.svg',
+      logo: '/logos/sonarqube.svg',
+      name: 'SonarQube',
+      description: 'test',
+    },
+    {
+      logo: '/logos/openai.svg',
       name: 'ChatGPT',
       description: 'test',
     },